+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.17
+
+- Fixed welcome screen.
+
 ## 1.0.16
 
 - Simplified welcome screen.